requirements.txt
Django==1.9.2
django-markup==0.4
django-ordered-model==1.1.0
django-pagedown==0.1.1
django-static-precompiler==1.3
django-suit==0.2.17
Markdown==2.6
Pillow==3.1.1
pyjade==4.0.0
six==1.10.0
wheel==0.29.0
